+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Preview of PlaceObject and ShowFrame in the Dump view
 - FileAttributes tag exception in the Dump view
 - Adding new frames did not set correct timelined to ShowFrame
+- Computing dependent characters inside DefineSprite
 
 ### Changed
 - [#1455] All tag types are now allowed inside DefineSprite